Refactor repository structure and organize files (Issue #4)
