+++ src/compression/datagen.py
